# openapi.yaml
openapi: 3.0.3
info:
  title: AIRS Backend API
  version: v2
  description: AIRS backend API definition for mobile/admin app and sensor ingestion query.
servers:
  - url: https://airs.bibnear.cloud
    description: Production API via Caddy HTTPS reverse proxy
  - url: http://localhost:8080
    description: Local development API
...
paths:
...
  /airs/admin/nodes:
    get:
      tags:
        - Admin Nodes
      summary: 관리자 노드 목록 조회
      description: 관리자 앱 노드 관리 화면을 위한 목록 API입니다. 승인된 ADMIN 또는 ROOT_ADMIN만 호출할 수 있습니다.
      security:
        - bearerAuth: []
      parameters:
        - name: sort
          in: query
          required: false
          description: 노드 목록 정렬 기준
          schema:
            type: string
            enum:
              - distance
              - status
              - space
              - alert
            default: distance
      responses:
        '200':
          description: 노드 목록 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminNodeListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /airs/admin/nodes/{nodeId}:
    get:
      tags:
        - Admin Nodes
      summary: 관리자 노드 상세 조회
      description: 관리자 앱 노드 상세 화면을 위한 API입니다. 승인된 ADMIN 또는 ROOT_ADMIN만 호출할 수 있습니다.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/NodeId'
      responses:
        '200':
          description: 노드 상세 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminNodeDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Admin Nodes
      summary: 관리자 노드 삭제
      description: 물리 노드 row를 삭제하지 않고 active installation을 비활성화합니다. 삭제한 노드에 연결된 ACTIVE 알림은 자동으로 RESOLVED 처리합니다.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/NodeId'
      responses:
        '204':
          description: 노드 삭제 성공
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /airs/admin/nodes/{nodeId}/co2-trend:
    get:
      tags:
        - Admin Nodes
      summary: 관리자 노드 CO2 추이 조회
      description: >
        InfluxDB의 telemetry co2_ppm 시계열을 집계하여 노드 상세 그래프에 반환합니다.
        호출 방식은 두 가지입니다. period=1d|5d|1mo|6mo|1y를 지정하면 기간별 기본 집계 간격을
        사용합니다(1d=10m, 5d=1h, 1mo=6h, 6mo=1d, 1y=1d). period를 생략하면 최근 hours시간을
        window 간격으로 집계합니다(기본 hours=6, window=5m). period와 hours/window를 함께 보내도
        오류가 나지 않으며, period가 우선되어 hours와 window는 무시됩니다. 예를 들어
        period=5d&hours=6&window=30s 요청은 최근 5일을 1시간 단위로 반환합니다. 6mo와 1y는
        완료된 날짜의 sensor_rollup_1d 집계를 우선 읽고, 오늘 또는 집계 공백은 raw telemetry로 보완합니다.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/NodeId'
        - name: period
          in: query
          required: false
          description: >
            프리셋 기간 모드입니다. 지정 시 hours/window보다 우선하며, 1d=10m, 5d=1h,
            1mo=6h, 6mo=1d, 1y=1d 기본 집계 간격을 사용합니다. 예: period=5d.
          schema:
            type: string
            enum:
              - 1d
              - 5d
              - 1mo
              - 6mo
              - 1y
            example: 5d
        - name: hours
          in: query
          required: false
          description: >
            사용자 정의 최근 시간 모드입니다. period를 생략했을 때만 사용하며 1~24만 허용합니다.
            window와 함께 사용합니다. 둘 다 생략하면 최근 6시간을 기본 5분 단위로 반환합니다.
          schema:
            type: integer
            minimum: 1
            maximum: 24
            default: 6
        - name: window
          in: query
          required: false
          description: >
            사용자 정의 최근 시간 모드의 Flux aggregateWindow 간격입니다. period를 생략했을 때만
            사용합니다. 예: hours=6&window=5m. 허용 형식은 숫자 뒤에 s, m, h, d가 붙는 값입니다.
          schema:
            type: string
            default: 5m
            example: 5m
      responses:
        '200':
          description: CO2 추이 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminNodeCo2TrendResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /airs/admin/nodes/{nodeId}/sensor-trend:
    get:
      tags:
        - Admin Nodes
      summary: 관리자 노드 선택 센서 추이 조회
      description: >
        노드 상세의 온도·습도·CO2 카드가 선택한 지표 하나만 반환합니다. metric과 period는 모두 필수입니다.
        1d는 raw sensor_data를 10분 평균으로 읽고, 5d는 sensor_rollup_1h의 1시간 평균을 우선 사용합니다.
        1mo는 1시간 rollup을 count 가중 6시간 평균으로 묶으며, 6mo와 1y는 sensor_rollup_1d의 일 평균을 우선 사용합니다.
        필요한 rollup이 없거나 날짜·시간 중간에 공백이 있거나 조회에 실패하면 같은 기간의 raw telemetry로 전체 fallback합니다.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/NodeId'
        - name: metric
          in: query
          required: true
          description: 그래프로 조회할 단일 센서 지표입니다.
          schema:
            type: string
            enum:
              - temperature
              - humidity
              - co2
            example: temperature
        - name: period
          in: query
          required: true
          description: >
            노드 상세의 고정 기간입니다. 1d=10m, 5d=1h, 1mo=6h, 6mo=1d, 1y=1d 집계 간격을 사용합니다.
          schema:
            type: string
            enum:
              - 1d
              - 5d
              - 1mo
              - 6mo
              - 1y
            example: 1mo
      responses:
        '200':
          description: 선택 센서 추이 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminNodeSensorTrendResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /airs/admin/analytics/co2/summary:
    get:
      tags:
        - Admin Analytics
      summary: 관리자 환기/CO2 요약 카드 조회
      description: 환기/CO2 분석 화면 상단의 환기 양호, 환기 권장, 환기 필요, 데이터 없음 카드만 빠르게 조회합니다. InfluxDB 시간대별 추이 조회는 수행하지 않습니다.
      security:
        - bearerAuth: []
      parameters:
        - name: date
          in: query
          required: false
          description: 조회 기준 날짜입니다. 생략하면 서버 기준 오늘 날짜를 사용합니다.
          schema:
            type: string
            format: date
            example: '2026-07-07'
      responses:
        '200':
          description: 환기/CO2 요약 카드 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminCo2SummaryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /airs/admin/analytics/co2/distribution:
    get:
      tags:
        - Admin Analytics
      summary: 관리자 CO2 평균 농도 분포 조회
      description: 환기/CO2 분석 화면의 CO2 평균 농도 분포만 빠르게 조회합니다. 응답 항목의 unit은 SPACE이며 totalCount는 분모로 사용한 전체 공간 수입니다.
      security:
        - bearerAuth: []
      parameters:
        - name: date
          in: query
          required: false
          description: 조회 기준 날짜입니다. 생략하면 서버 기준 오늘 날짜를 사용합니다.
          schema:
            type: string
            format: date
            example: '2026-07-07'
      responses:
        '200':
          description: CO2 평균 농도 분포 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminCo2DistributionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /airs/admin/analytics/co2/trend:
    get:
      tags:
        - Admin Analytics
      summary: 관리자 오늘/어제 CO2 시간대별 평균 추이 조회
      description: 분석 요약과 환기/CO2 분석 화면이 공통으로 사용하는 오늘/어제 CO2 평균 추이 그래프를 한 번에 조회합니다. 서버는 기준일 전날부터 기준일 종료까지 48시간 범위를 InfluxDB에 한 번 조회한 뒤, Asia/Seoul 날짜 경계로 todayTrend와 yesterdayTrend를 분리합니다. 두 배열은 각각 1시간 단위 평균입니다.
      security:
        - bearerAuth: []
      parameters:
        - name: date
          in: query
          required: false
          description: 조회 기준 날짜입니다. 생략하면 서버 기준 오늘 날짜를 사용합니다.
          schema:
            type: string
            format: date
            example: '2026-07-07'
      responses:
        '200':
          description: 환기/CO2 시간대별 평균 추이 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminCo2TrendResponse'
              example:
                campusId: 1
                campusName: 서강대학교
                date: '2026-07-07'
                todayTrend:
                  - timestamp: '2026-07-07T00:00:00Z'
                    co2Ppm: 612
                  - timestamp: '2026-07-07T01:00:00Z'
                    co2Ppm: 680
                  - timestamp: '2026-07-07T02:00:00Z'
                    co2Ppm: 842
                yesterdayTrend:
                  - timestamp: '2026-07-06T00:00:00Z'
                    co2Ppm: 590
                  - timestamp: '2026-07-06T01:00:00Z'
                    co2Ppm: 640
                  - timestamp: '2026-07-06T02:00:00Z'
                    co2Ppm: 700
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /airs/admin/analytics/co2/top-spaces:
    get:
      tags:
        - Admin Analytics
      summary: 관리자 CO2 Top 5 공간 조회
      description: 환기/CO2 분석 화면의 CO2 Top 5 공간 목록만 빠르게 조회합니다. InfluxDB 시간대별 추이 조회는 수행하지 않습니다.
      security:
        - bearerAuth: []
      parameters:
        - name: date
          in: query
          required: false
          description: 조회 기준 날짜입니다. 생략하면 서버 기준 오늘 날짜를 사용합니다.
          schema:
            type: string
            format: date
            example: '2026-07-07'
      responses:
        '200':
          description: CO2 Top 5 공간 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminCo2TopSpacesResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /airs/admin/analytics/overview/status-distributions:
    get:
      tags:
        - Admin Analytics
      summary: 관리자 분석 요약 상태 분포 조회
      description: 분석 요약 화면의 CO2, 연결, 재실, Wi-Fi 상태 분포만 빠르게 조회합니다. CO2/재실은 공간 기준, 연결/Wi-Fi는 노드 기준이며 각 항목의 unit과 totalCount를 함께 반환합니다. InfluxDB 시간대별 추이 조회는 수행하지 않습니다.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 상태 분포 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminAnalyticsStatusDistributionsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /airs/admin/nodes/installations:
    post:
      tags:
        - Admin Nodes
      summary: Bluetooth 기반 노드 등록 완료
      description: 앱이 Bluetooth/Wi-Fi 설정 성공을 확인한 뒤 backend에 최종 설치 결과를 저장합니다. Wi-Fi password는 절대 보내지 않습니다.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminNodeRegistrationRequest'
      responses:
        '201':
          description: 새 노드 설치 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminNodeRegistrationResponse'
        '200':
          description: 같은 nodeId와 spaceId의 중복 등록 요청. 기존 설치 결과를 반환합니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminNodeRegistrationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Authorization: Bearer {accessToken}'
  parameters:
    NodeId:
      name: nodeId
      in: path
      required: true
      schema:
        type: string
        example: node_01
...
  responses:
    BadRequest:
      description: 요청 값이 올바르지 않습니다.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Unauthorized:
      description: 인증이 필요하거나 토큰이 유효하지 않습니다.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Forbidden:
      description: 접근 권한이 없습니다.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    NotFound:
      description: 요청한 리소스를 찾을 수 없습니다.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Conflict:
      description: 리소스 상태가 충돌합니다.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
  schemas:
...
    ConnectionStatus:
      type: string
      enum:
        - ONLINE
        - WEAK
        - OFFLINE
        - UNKNOWN
    OccupancyStatus:
      type: string
      enum:
        - OCCUPIED
        - UNOCCUPIED
        - UNKNOWN
...
    AlertSeverity:
      type: string
      enum:
        - EMERGENCY
        - WARNING
        - INFO
...
    AdminNodeListResponse:
      type: object
      properties:
        campusId:
          type: integer
          format: int64
        campusName:
          type: string
          example: 서강대학교
        radiusMeter:
          type: integer
          nullable: true
          example: 500
        totalNodeCount:
          type: integer
        onlineNodeCount:
          type: integer
          format: int64
        weakNodeCount:
          type: integer
          format: int64
        offlineNodeCount:
          type: integer
          format: int64
        activeAlertCount:
          type: integer
          format: int64
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/AdminNodeSummaryResponse'
    AdminNodeSummaryResponse:
      type: object
      properties:
        rank:
          type: integer
          example: 1
        nodeId:
          type: string
          example: node_01
        spaceId:
          type: integer
          format: int64
          example: 1
        spaceCode:
          type: string
          example: K301
        spaceName:
          type: string
          example: 301호
        buildingName:
          type: string
          example: 김대건관
        floorLabel:
          type: string
          nullable: true
          example: 3층
        distanceMeter:
          type: integer
          nullable: true
        connectionStatus:
          $ref: '#/components/schemas/ConnectionStatus'
        temperature:
          type: number
          nullable: true
          example: 24.3
        humidity:
          type: number
          nullable: true
          example: 52.0
        co2Ppm:
          type: integer
          nullable: true
          example: 842
        co2Summary:
          type: string
          nullable: true
          example: 보통
        occupancyStatus:
          allOf:
            - $ref: '#/components/schemas/OccupancyStatus'
          nullable: true
        occupancySummary:
          type: string
          nullable: true
          example: 있음
        alertCount:
          type: integer
          format: int64
        lastUpdatedAt:
          type: string
          format: date-time
          nullable: true
    AdminNodeDetailResponse:
      type: object
      properties:
        nodeId:
          type: string
          example: node_01
        spaceId:
          type: integer
          format: int64
        spaceCode:
          type: string
          example: K301
        spaceName:
          type: string
          example: 301호
        buildingName:
          type: string
          example: 김대건관
        floorLabel:
          type: string
          nullable: true
          example: 3층
        firmwareVersion:
          type: string
          nullable: true
          example: v1.2.3
        installedAt:
          type: string
          format: date-time
          nullable: true
        lastUpdatedAt:
          type: string
          format: date-time
          nullable: true
        connectionStatus:
          $ref: '#/components/schemas/ConnectionStatus'
        wifiRssi:
          type: integer
          nullable: true
          example: -45
        wifiRssiSummary:
          type: string
          nullable: true
          example: 강함
        temperature:
          type: number
          nullable: true
          example: 24.3
        temperatureSummary:
          type: string
          nullable: true
          example: 보통
        humidity:
          type: number
          nullable: true
          example: 52.0
        humiditySummary:
          type: string
          nullable: true
          example: 보통
        co2Ppm:
          type: integer
          nullable: true
          example: 842
        co2Summary:
          type: string
          nullable: true
          example: 보통
        humanDetected:
          type: boolean
          nullable: true
        occupancyStatus:
          allOf:
            - $ref: '#/components/schemas/OccupancyStatus'
          nullable: true
        occupancySummary:
          type: string
          nullable: true
          example: 현재 재실
        comfortScore:
          type: number
          nullable: true
          example: 82.0
        comfortSummary:
          type: string
          nullable: true
          example: 쾌적
        alerts:
          type: array
          items:
            $ref: '#/components/schemas/AdminNodeAlertResponse'
    AdminNodeAlertResponse:
      type: object
      properties:
        alertId:
          type: integer
          format: int64
        alertType:
          $ref: '#/components/schemas/AlertType'
        severity:
          $ref: '#/components/schemas/AlertSeverity'
        title:
          type: string
          example: 환기 권장
        message:
          type: string
          nullable: true
          example: 최근 30분 동안 CO2가 상승했습니다.
        lastDetectedAt:
          type: string
          format: date-time
    AdminNodeRegistrationRequest:
      type: object
      required:
        - nodeId
        - spaceId
      properties:
        nodeId:
          type: string
          maxLength: 80
          example: node_01
          description: MQTT topic의 실제 nodeId와 동일하게 저장합니다.
        spaceId:
          type: integer
          format: int64
          example: 1
        hardwareVersion:
          type: string
          maxLength: 50
          nullable: true
          example: ESP32-C3
        firmwareVersion:
          type: string
          maxLength: 50
          nullable: true
          example: v1.2.3
        wifiRssi:
          type: integer
          minimum: -120
          maximum: 0
          nullable: true
          example: -45
    AdminNodeRegistrationResponse:
      type: object
      properties:
        nodeId:
          type: string
          example: node_01
        spaceId:
          type: integer
          format: int64
        spaceCode:
          type: string
          example: K301
        spaceName:
          type: string
          example: 301호
        buildingName:
          type: string
          example: 김대건관
        installedAt:
          type: string
          format: date-time
        created:
          type: boolean
          description: true면 새 등록, false면 같은 nodeId와 spaceId의 중복 등록 요청입니다.
    AdminNodeCo2TrendResponse:
      type: object
      properties:
        nodeId:
          type: string
          example: node_01
        period:
          type: string
          nullable: true
          description: period 기반 조회일 때 선택된 기간입니다. 기존 hours/window 조회이면 null입니다.
          enum:
            - 1d
            - 5d
            - 1mo
            - 6mo
            - 1y
          example: 5d
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
        window:
          type: string
          example: 5m
        points:
          type: array
          items:
            $ref: '#/components/schemas/AdminNodeCo2TrendPointResponse'
    AdminNodeCo2TrendPointResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        co2Ppm:
          type: integer
          nullable: true
    AdminNodeSensorTrendResponse:
      type: object
      required:
        - nodeId
        - metric
        - period
        - from
        - to
        - window
        - points
      properties:
        nodeId:
          type: string
          example: node_01
        metric:
          type: string
          enum:
            - temperature
            - humidity
            - co2
          example: temperature
        period:
          type: string
          enum:
            - 1d
            - 5d
            - 1mo
            - 6mo
            - 1y
          example: 1d
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
        window:
          type: string
          example: 10m
        points:
          type: array
          items:
            $ref: '#/components/schemas/AdminNodeSensorTrendPointResponse'
...
    AdminCo2SummaryResponse:
      type: object
      properties:
        campusId:
          type: integer
          format: int64
        campusName:
          type: string
          example: 서강대학교
        date:
          type: string
          format: date
          example: '2026-07-07'
        totalSpaceCount:
          type: integer
          example: 25
        averageCo2Ppm:
          type: integer
          nullable: true
          example: 842
        ventilationSummary:
          $ref: '#/components/schemas/AdminCo2VentilationSummaryResponse'
    AdminCo2DistributionResponse:
      type: object
      properties:
        campusId:
          type: integer
          format: int64
        campusName:
          type: string
          example: 서강대학교
        date:
          type: string
          format: date
          example: '2026-07-07'
        totalSpaceCount:
          type: integer
          example: 25
        averageCo2Ppm:
          type: integer
          nullable: true
          example: 842
        distribution:
          type: array
          items:
            $ref: '#/components/schemas/AdminCo2DistributionItemResponse'
    AdminCo2TrendResponse:
      type: object
      properties:
        campusId:
          type: integer
          format: int64
        campusName:
          type: string
          example: 서강대학교
        date:
          type: string
          format: date
          example: '2026-07-07'
        todayTrend:
          type: array
          description: 조회 기준 날짜의 시간대별 CO2 평균 배열입니다.
          items:
            $ref: '#/components/schemas/AdminCo2TrendPointResponse'
        yesterdayTrend:
          type: array
          description: 조회 기준 날짜 하루 전의 시간대별 CO2 평균 배열입니다.
          items:
            $ref: '#/components/schemas/AdminCo2TrendPointResponse'
    AdminCo2TopSpacesResponse:
      type: object
      properties:
        campusId:
          type: integer
          format: int64
        campusName:
          type: string
          example: 서강대학교
        date:
          type: string
          format: date
          example: '2026-07-07'
        topSpaces:
          type: array
          items:
            $ref: '#/components/schemas/AdminCo2TopSpaceResponse'
...
    AdminAnalyticsStatusDistributionsResponse:
      type: object
      properties:
        co2:
          type: array
          items:
            $ref: '#/components/schemas/AdminAnalyticsDistributionItemResponse'
        connection:
          type: array
          items:
            $ref: '#/components/schemas/AdminAnalyticsDistributionItemResponse'
        occupancy:
          type: array
          items:
            $ref: '#/components/schemas/AdminAnalyticsDistributionItemResponse'
        wifi:
          type: array
          items:
            $ref: '#/components/schemas/AdminAnalyticsDistributionItemResponse'
...
    ApiErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string